requirements.txt
django
django-ninja
psycopg2-binary
psycopg
pgvector
langchain
langchain-community
langchain-postgres
langchain-text-splitters
pdfplumber
openai
python-dotenv
langchain-openai
transformers
torch
edge-tts
